requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Cache & Sessions
redis==5.0.1
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client & WebSockets
httpx==0.25.2
websockets==12.0
python-socketio==5.10.0

# Monitoring & Metrics
prometheus-client==0.19.0

# Task Queue (Optional for demo)
celery==5.3.4
flower==2.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional Utilities
python-dateutil==2.8.2
pytz==2023.3

# NATS Integration (Optional - falls back to logging if not available)
nats-py==2.6.0